Added webhose.io api script
This should pull in results from all news sites webhose.io indexes for the first 44 detention centres in our spreadsheet. We need to create separate calls for each set of (approx) 44, because of restrictions on how long queries can be. There is also a non-script way to implement this and return results as spreadsheet (or rss, or json or whatever you like) via the webhose.io site: https://webhose.io/web-content-api)