Added detention_centres_query.txt
This file just contains a text string with all the detention centres from the spreadsheet (not including alternative names), joined with " OR " and with special characters escaped. For use in api queries. (You may have to use a subset at a time so as not to overload the character limit on qs).  It was created by creating a new column in the spreadsheet to the right of C and using =""""&C2&""""&" OR "&""""&C3&"""" in row 2 and =D2&" OR " &""""&C2&"""" in row 3. Then grabbing the output at the bottom of this column, saving the values only, and using regex to replace special characters with escaped versions.